+++ certificate/letsencrypt/dns/connbyte/provider.go
@@ -1,0 +1,55 @@
+package connbyte
+
+import (
+	"context"
+
+	"github.com/go-acme/lego/v5/challenge"
+	"github.com/go-acme/lego/v5/providers/dns/connbyte"
+
+	"dillmann.com.br/nginx-ignition/certificate/letsencrypt/dns"
+	"dillmann.com.br/nginx-ignition/core/common/dynamicfields"
+	"dillmann.com.br/nginx-ignition/core/common/i18n"
+)
+
+//nolint:gosec
+const (
+	tokenFieldID = "connbyteToken"
+)
+
+type Provider struct{}
+
+func (p *Provider) ID() string {
+	return "CONNBYTE"
+}
+
+func (p *Provider) Name(ctx context.Context) *i18n.Message {
+	return i18n.M(ctx, i18n.K.CertificateLetsencryptDnsConnbyteName)
+}
+
+func (p *Provider) DynamicFields(ctx context.Context) []dynamicfields.DynamicField {
+	return dns.LinkedToProvider(p.ID(), []dynamicfields.DynamicField{
+		{
+			ID:          tokenFieldID,
+			Description: i18n.M(ctx, i18n.K.CertificateLetsencryptDnsConnbyteToken),
+			Required:    true,
+			Sensitive:   true,
+			Type:        dynamicfields.SingleLineTextType,
+		},
+	})
+}
+
+func (p *Provider) ChallengeProvider(
+	_ context.Context,
+	_ []string,
+	parameters map[string]any,
+) (challenge.Provider, error) {
+	token, _ := parameters[tokenFieldID].(string)
+
+	cfg := connbyte.NewDefaultConfig()
+	cfg.Token = token
+	cfg.PropagationTimeout = dns.PropagationTimeout
+	cfg.PollingInterval = dns.PollingInterval
+	cfg.TTL = dns.TTL
+
+	return connbyte.NewDNSProviderConfig(cfg)
+}