+++ certificate/letsencrypt/dns/dynadot/provider.go
@@ -1,0 +1,64 @@
+package dynadot
+
+import (
+	"context"
+
+	"github.com/go-acme/lego/v5/challenge"
+	"github.com/go-acme/lego/v5/providers/dns/dynadot"
+
+	"dillmann.com.br/nginx-ignition/certificate/letsencrypt/dns"
+	"dillmann.com.br/nginx-ignition/core/common/dynamicfields"
+	"dillmann.com.br/nginx-ignition/core/common/i18n"
+)
+
+//nolint:gosec
+const (
+	apiKeyFieldID    = "dynadotApiKey"
+	apiSecretFieldID = "dynadotApiSecret"
+)
+
+type Provider struct{}
+
+func (p *Provider) ID() string {
+	return "DYNADOT"
+}
+
+func (p *Provider) Name(ctx context.Context) *i18n.Message {
+	return i18n.M(ctx, i18n.K.CertificateLetsencryptDnsDynadotName)
+}
+
+func (p *Provider) DynamicFields(ctx context.Context) []dynamicfields.DynamicField {
+	return dns.LinkedToProvider(p.ID(), []dynamicfields.DynamicField{
+		{
+			ID:          apiKeyFieldID,
+			Description: i18n.M(ctx, i18n.K.CertificateLetsencryptDnsDynadotApiKey),
+			Required:    true,
+			Type:        dynamicfields.SingleLineTextType,
+		},
+		{
+			ID:          apiSecretFieldID,
+			Description: i18n.M(ctx, i18n.K.CertificateLetsencryptDnsDynadotApiSecret),
+			Required:    true,
+			Sensitive:   true,
+			Type:        dynamicfields.SingleLineTextType,
+		},
+	})
+}
+
+func (p *Provider) ChallengeProvider(
+	_ context.Context,
+	_ []string,
+	parameters map[string]any,
+) (challenge.Provider, error) {
+	apiKey, _ := parameters[apiKeyFieldID].(string)
+	apiSecret, _ := parameters[apiSecretFieldID].(string)
+
+	cfg := dynadot.NewDefaultConfig()
+	cfg.APIKey = apiKey
+	cfg.APISecret = apiSecret
+	cfg.PropagationTimeout = dns.PropagationTimeout
+	cfg.PollingInterval = dns.PollingInterval
+	cfg.TTL = dns.TTL
+
+	return dynadot.NewDNSProviderConfig(cfg)
+}