+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 2.39.0
 
 - Optimization of the Docker image to use the latest nginx mainline version while reducing its size and layers.
+- Added support for 2 new DNS providers for Let's Encrypt challenges: Connbyte and Dynadot.
+- Removed the StackPath DNS provider as it shut down in 2024.
 
 ## 2.38.0